Allow more natural updating of values
A number of the QuickCheck tests indicated that there's a potential
inconsistency around how we're modelling the queries. For example should
`cat` have been parsed as `Cat []` or `Cat [Query []]`.

This is causing a ripple effect out through the code so the diff here is
quite large (and will get larger for a while), but I expect it to
eventually make the code more straight forward.

We also find that we weren't testing `Mars.Eval`, so adding a simple
test for that now, there'll almost certainly be a refactoring (Free
Monad? 🤔) to improve that story too.