Allow testing of command effects
Many of the interesting effects of the commands when evaluated are
strictly pure other than they might output to the terminal (`save` and
`load` being the only current examples which actually have other effects in
the outside world)

This observation means that we can move most of their logic outside
of the IO monad and just have an `eval` function responsible for
printing the output.

This makes it much easier to start testing the functionality of these
operations in unit tests. Testing `ls` for instance found that the
behaviour was currently correct, and in fact that the code could be a
lot simpler. ♥ Haskell