Extract commands into individual types
The `Command` type is getting unwieldy, so moving the constituent parts
out into individual types. This was intended to include parsing, but we
quickly run into the Expression Problem as we need to have a specific
type to return (not just a typeclass). There's a couple of potential
type level tricks that could resolve this, but we can come back to that
later (they may add complexity which we don't need).