Extract commands into individual types (#6)
* Extract commands into individual types

The `Command` type is getting unwieldy, so moving the constituent parts
out into individual types. This was intended to include parsing, but we
quickly run into the Expression Problem as we need to have a specific
type to return (not just a typeclass). There's a couple of potential
type level tricks that could resolve this, but we can come back to that
later (they may add complexity which we don't need).

* Remove unnecessary `readCommand` from Command typeclass

* Hlint suggestion

* Fix Client to work with new code structure

* Dynamically define executable for script/dev

* Hlint suggestions

* Restore evalCommand and printCommand for Commands

* Don't import Mars.Eval in tests