Split out `Command` and `Action` typeclasses
The Command typeclass combined both the pure and the state changing
code into one concept, which resulted in some places where there was a
pretty tight coupling of the result of an evaluation and it's final
representation on screen¹.

We're separating these concepts into `Command`, which takes a state and
a command and returns an event describing the result of this command,
and an Action which takes this result and applies it to the state
(potentially including side effects and IO).

¹ `ls` being a particular example, where the first pass at a test for it
required that the test include the concept of green text, when all it
needed to account for was the name of the objects it was listing.