Correctly parse `..` as a query for commands
The implementation of `parser` was not correctly parsing commands like
`cd ..` and `ls ..`.

The reason was that the `commandLine` parser was expecting a list of
commands separated by `|`, for future functionality allowing piping
between commands. I'm somewhat unclear as to why it's being so eager to
consume this input.

However this falls into YAGNI territory as we don't currently have an
commands that can usefully operate on the output of another command.
Therefore we will remove it for now, and can revisit if/when this turns
out to be useful